Modificacion de la clase que guardara con el one to many de modelos

+++ src/main/java/com/udemy/service/impl/ModeloServiceImpl.java
@@ -1,13 +1,11 @@
 package com.udemy.service.impl;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.udemy.model.Modelo;
 import com.udemy.repo.IModeloRepository;
+import com.udemy.repo.crude.IModeloCrudeRespository;
 import com.udemy.service.IModeloService;
 import com.udemy.service.common.ServiceCommon;
 
@@ -17,27 +15,21 @@ public class ModeloServiceImpl extends ServiceCommon<Modelo> implements IModeloS
 	@Autowired
 	IModeloRepository repo;
 
-	private List<Modelo> modelos;
-
-	public ModeloServiceImpl() {
-		modelos = new ArrayList<Modelo>();
-	}
+	@Autowired
+	IModeloCrudeRespository crudeRepo;
 
 	@Override
 	public int eliminar(Integer id) {
 		try {
+
+			if (GetById(id) == null)
+				return 0;
+
 			repo.deleteById(id);
 			return 1;
 		} catch (Exception e) {
 			throw e;
 		}
-	}
-	
-	@Override
-	public int eliminarModeloFromList(String codigo) {
-		if (modelos.removeIf(modelo -> modelo.getCodigo() == codigo))
-			return 1;
-		return -1;
 	}
 
 	@Override
@@ -51,17 +43,16 @@ public class ModeloServiceImpl extends ServiceCommon<Modelo> implements IModeloS
 	}
 
 	@Override
-	public int agergar(Modelo modelo) {
+	public int currentId() {
+		return crudeRepo.max();
+	}
+
+	private Modelo GetById(Integer id) {
 		try {
-			modelos.add(modelo);
-			return 1;
+			object = repo.findById(id);
+			return object.isPresent() == true ? object.get() : null;
 		} catch (Exception e) {
 			throw e;
 		}
-	}
-
-	@Override
-	public List<Modelo> getModelos() {
-		return modelos;
 	}
 }
